refactor(core): rename turn_state_machine.rs to turn.rs
Shorter module name since all types already have Turn prefix.
Usage: `use crate::turn::{TurnContext, TurnStateMachine}`
