+++ crates/core/src/lib.rs
@@ -69,8 +69,8 @@ pub mod llmsim_driver;
 // In-memory agentic loop for testing and prototyping
 pub mod in_memory_loop;
 
-// Turn state machine for unified turn orchestration
-pub mod turn_state_machine;
+// Turn orchestration (state machine, context, outcomes)
+pub mod turn;
 
 // Note: LLM Driver implementations (AnthropicLlmDriver, OpenAILlmDriver) are now in
 // separate crates (everruns-anthropic, everruns-openai) that depend on everruns-core.
